# docs/openapi/tenant-openapi.yaml
openapi: 3.1.0
info:
  title: CloudAPI Tenant
  description: '

    Welcome to CloudAPI Tenant!


    For detailed guidance on using the API, please visit our official documentation:

    https://www.didx.co.za/acapy-cloud/index.html

'
  version: 5.0.0-rc5
servers:
  - url: /tenant
paths:
...
  /v1/wallet/sd-jws/sign:
    post:
      tags:
        - wallet
      summary: Sign SD-JWS
      description: "Sign a Selective Disclosure JSON Web Signature (SD-JWS).\n---\n\nThis endpoint allows users to create a Selective Disclosure JSON Web Signature (SD-JWS).\nThe SD-JWS enables the selective disclosure of specific attributes to a verifier while keeping others confidential.\n\n**Usage:**\n\n- **DID-Based Signing:** Provide the `did` field with a valid DID.\nThe Aries agent will automatically select the appropriate verification key associated with the DID.\n\n- **Verification Method-Based Signing:** Provide the `verification_method` field with a specific verification method\n(DID with verkey) to explicitly specify which key to use for signing.\n\n**Notes:**\n\n- The `headers` field is optional. Custom headers can be specified, but the `typ`, `alg`,\n  and `kid` fields are automatically populated by the Aries agent based on the signing method.\n- The `non_sd_list` field specifies attributes that are **not** selectively disclosed.\n  Attributes listed here will always be included in the SD-JWS.\n\n**Non-Selective Disclosure (`non_sd_list`):**\n\n- To exclude list elements:\n    - Use the format `\"<attribute_name>[start:end]\"` where `start` and `end` define the range\n      (e.g., `\"nationalities[1:3]\"`).\n- To exclude specific dictionary attributes:\n    - Use the format `\"<dictionary_name>.<attribute_name>\"` (e.g., `\"address.street_address\"`).\n\n**Example Request Body:**\n```json\n{\n    \"did\": \"did:cheqd:...\",\n    \"payload\": {\n        \"credential_subject\": \"reference_to_holder\",\n        \"given_name\": \"John\",\n        \"family_name\": \"Doe\",\n        \"email\": \"johndoe@example.com\",\n        \"phone_number\": \"+27-123-4567\",\n        \"nationalities\": [\"a\",\"b\",\"c\",\"d\"],\n        \"address\": {\n            \"street_address\": \"123 Main St\",\n            \"locality\": \"Anytown\",\n            \"region\": \"Anywhere\",\n            \"country\": \"ZA\"\n        },\n        \"birthdate\": \"1940-01-01\"\n    },\n    \"non_sd_list\": [\n        \"given_name\",\n        \"address\",\n        \"address.street_address\",\n        \"nationalities\",\n        \"nationalities[1:3]\"\n    ]\n}\n```\n\nRequest Body:\n---\n    SDJWSCreateRequest:\n        `did` (str, optional): The DID to sign the SD-JWS with.\n        `verification_method` (str, optional): The verification method (DID with verkey) to use for signing.\n        `payload` (dict): The JSON payload to be signed.\n        `headers` (dict, optional): Custom headers for the SD-JWS.\n        `non_sd_list` (List[str], optional): List of attributes excluded from selective disclosure.\n\nResponse:\n---\n    SDJWSCreateResponse:\n        `sd_jws` (str): The resulting SD-JWS string concatenated with the necessary disclosures in the format\n        `<Issuer-signed JWS>~<Disclosure 1>~<Disclosure 2>~...~<Disclosure N>`.\n\n**References:**\n\n- [Selective Disclosure JSON Web Token (SD-JWT)\n  Specification](https://www.ietf.org/archive/id/draft-ietf-oauth-selective-disclosure-jwt-07.html)"
      operationId: sign_sd_jws_v1_wallet_sd_jws_sign_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SDJWSCreateRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SDJWSCreateResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - APIKeyHeader: []
  /v1/wallet/sd-jws/verify:
    post:
      tags:
        - wallet
      summary: Verify SD-JWS
      description: "Verify a Selective Disclosure JSON Web Signature (SD-JWS).\n---\n\nThis endpoint allows users to verify the authenticity and integrity of a Selective Disclosure\nJSON Web Signature (SD-JWS). It decodes the SD-JWS to retrieve the payload and headers,\nassesses its validity, and processes the disclosures.\n\n**Usage:**\n\n- Submit the SD-JWS string concatenated with the necessary disclosures to this endpoint.\n- The format should be: `<Issuer-signed JWS>~<Disclosure 1>~<Disclosure 2>~...~<Disclosure N>`.\n- The holder provides the SD-JWS along with the required disclosures based on the verifier's request.\n\n**Notes:**\n\n- Only the disclosures relevant to the verifier's request needs to be provided.\n  Other disclosures can remain confidential.\n\n**Example Request Body:**\n```json\n{\n    \"sd_jws\": \"<Issuer-signed JWS>~<Disclosure 1>~<Disclosure 2>~...~<Disclosure N>\"\n}\n```\n\nRequest Body:\n---\n    SDJWSVerifyRequest:\n        `sd_jws` (str): The concatenated SD-JWS and disclosures to verify and reveal.\n\nResponse:\n---\n    SDJWSVerifyResponse:\n        `valid` (bool): Indicates whether the SD-JWS is valid.\n        `payload` (dict): The decoded payload of the SD-JWS.\n        `headers` (dict): The headers extracted from the SD-JWS.\n        `kid` (str): The Key ID of the signer.\n        `disclosed_attributes` (dict): The selectively disclosed attributes based on the provided disclosures.\n        `error` (str, optional): Error message if the SD-JWS verification fails.\n\n**References:**\n\n- [Selective Disclosure JSON Web Token (SD-JWT)\n  Specification](https://www.ietf.org/archive/id/draft-ietf-oauth-selective-disclosure-jwt-07.html)"
      operationId: verify_sd_jws_v1_wallet_sd_jws_verify_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SDJWSVerifyRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SDJWSVerifyResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - APIKeyHeader: []
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    SDJWSCreateRequest:
      properties:
        did:
          anyOf:
            - type: string
            - type: 'null'
          title: Did
          examples:
            - did:key:z6MkjCjxuTXxVPWS9JYj2ZiKtKvSS1srC6kBRes4WCB2mSWq
        headers:
          additionalProperties: true
          type: object
          title: Headers
          default: {}
        payload:
          additionalProperties: true
          type: object
          title: Payload
          description: Payload to sign
        verification_method:
          anyOf:
            - type: string
            - type: 'null'
          title: Verification Method
          description: Information used for proof verification
          examples:
            - did:key:z6MkjCjxuTXxVPWS9JYj2ZiKtKvSS1srC6kBRes4WCB2mSWq#z6MkjCjxuTXxVPWS9JYj2ZiKtKvSS1srC6kBRes4WCB2mSWq
        non_sd_list:
          anyOf:
            - items:
                type: string
              type: array
            - type: 'null'
          title: Non Sd List
          default: []
          examples:
            - - name
              - address
              - address.street_address
              - nationalities[1:3]
      type: object
      required:
        - payload
      title: SDJWSCreateRequest
    SDJWSCreateResponse:
      properties:
        sd_jws:
          type: string
          title: Sd Jws
          examples:
            - eyJhbGciOiJFZERTQSJ9.eyJhIjogIjAifQ.dBjftJeZ4CVP-mB92K27uhbUJU1p1r_wW1gFWFOEjXk~WyJEM3BUSFdCYWNRcFdpREc2TWZKLUZnIiwgIkRFIl0~WyJPMTFySVRjRTdHcXExYW9oRkd0aDh3IiwgIlNBIl0~WyJkVmEzX1JlTGNsWTU0R1FHZm5oWlRnIiwgInVwZGF0ZWRfYXQiLCAxNTcwMDAwMDAwXQ
      type: object
      required:
        - sd_jws
      title: SDJWSCreateResponse
    SDJWSVerifyRequest:
      properties:
        sd_jws:
          type: string
          title: Sd Jws
          examples:
            - eyJhbGciOiJFZERTQSJ9.eyJhIjogIjAifQ.dBjftJeZ4CVP-mB92K27uhbUJU1p1r_wW1gFWFOEjXk~WyJEM3BUSFdCYWNRcFdpREc2TWZKLUZnIiwgIkRFIl0~WyJPMTFySVRjRTdHcXExYW9oRkd0aDh3IiwgIlNBIl0~WyJkVmEzX1JlTGNsWTU0R1FHZm5oWlRnIiwgInVwZGF0ZWRfYXQiLCAxNTcwMDAwMDAwXQ
      type: object
      required:
        - sd_jws
      title: SDJWSVerifyRequest
    SDJWSVerifyResponse:
      properties:
        error:
          anyOf:
            - type: string
            - type: 'null'
          title: Error
          description: Error text
        headers:
          additionalProperties: true
          type: object
          title: Headers
          description: Headers from verified JWT.
          examples:
            - alg: EdDSA
              kid: did:key:z6MkjCjxuTXxVPWS9JYj2ZiKtKvSS1srC6kBRes4WCB2mSWq#z6MkjCjxuTXxVPWS9JYj2ZiKtKvSS1srC6kBRes4WCB2mSWq
              typ: JWT
        kid:
          type: string
          title: Kid
          description: kid of signer
          examples:
            - did:key:z6MkjCjxuTXxVPWS9JYj2ZiKtKvSS1srC6kBRes4WCB2mSWq#z6MkjCjxuTXxVPWS9JYj2ZiKtKvSS1srC6kBRes4WCB2mSWq
        payload:
          additionalProperties: true
          type: object
          title: Payload
          description: Payload from verified JWT
        valid:
          type: boolean
          title: Valid
        disclosures:
          items:
            items:
              anyOf:
                - type: string
                - additionalProperties: true
                  type: object
            type: array
          type: array
          title: Disclosures
          description: Disclosure arrays associated with the SD-JWT
          examples:
            - - - fx1iT_mETjGiC-JzRARnVg
                - name
                - Alice
              - - n4-t3mlh8jSS6yMIT7QHnA
                - street_address
                - _sd:
                    - kLZrLK7enwfqeOzJ9-Ss88YS3mhjOAEk9lr_ix2Heng
      type: object
      required:
        - headers
        - kid
        - payload
        - valid
        - disclosures
      title: SDJWSVerifyResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError
  securitySchemes:
    APIKeyHeader:
      type: apiKey
      in: header
      name: x-api-key